# engine/nodejs-dashboard/src/pages/openApi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OntoWebPT - Backend API",
    "version": "1.0.0",
    "description": "API for ingesting, listing and retrieving HTTP request data, analyzer findings, techstack findings, SPARQL access and PCAP processing."
  },
  "paths": {
...
    "/techstack/analyze": {
      "post": {
        "tags": ["Techstack"],
        "summary": "Analyze detected technologies, WAF, headers and cookies",
        "description": "Enqueues a job to analyze the technology stack based on headers, cookies, WAF hints and client-detected technologies.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechstackAnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Techstack analysis job accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accepted": { "type": "boolean" },
                    "jobId": { "type": "string" },
                    "technologies": {
                      "type": "integer",
                      "description": "Number of technologies in the payload"
                    },
                    "waf": {
                      "type": "integer",
                      "description": "Number of WAF entries in the payload"
                    },
                    "secureHeaders": {
                      "type": "integer",
                      "description": "Number of secureHeaders entries in the payload"
                    },
                    "cookies": {
                      "type": "integer",
                      "description": "Number of cookies in the payload"
                    }
                  },
                  "required": [
                    "accepted",
                    "jobId",
                    "technologies",
                    "waf",
                    "secureHeaders",
                    "cookies"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Enqueue failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/techstack/results/{jobId}": {
      "get": {
        "tags": ["Techstack"],
        "summary": "Get status or result of a techstack analysis job",
        "description": "Returns job execution status and analyzed technology stack details.",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" },
                    "jobId": { "type": "string" }
                  },
                  "required": ["error", "jobId"]
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/techstack/finding/list": {
      "get": {
        "tags": ["Techstack"],
        "summary": "List techstack findings (paginated)",
        "description": "Returns findings detected by the Techstack resolver, without detailed evidence.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 0, "maximum": 1000, "default": 100 }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 0, "maximum": 1000000, "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Findings list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechstackFindingListResponse"
                }
              }
            }
          },
          "502": {
            "description": "GraphDB query failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/techstack/finding/{id}": {
      "get": {
        "tags": ["Techstack"],
        "summary": "Get techstack finding details",
        "description": "Returns complete detail for a single Techstack finding, including software evidence, header evidence and cookie evidence.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Finding detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechstackFindingDetail"
                }
              }
            }
          },
          "404": {
            "description": "Finding not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "GraphDB query failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "detail": { "type": "string" }
        },
        "required": ["error"],
        "additionalProperties": true
      },
...
      "JobStatus": {
        "type": "object",
        "properties": {
          "jobId": { "type": "string" },
          "state": {
            "type": "string",
            "description": "Queue job state (e.g. waiting, active, completed, failed)"
          },
          "result": {
            "type": "object",
            "nullable": true,
            "description": "Job result payload, if available; structure depends on job type.",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "description": "Job creation timestamp (ms since epoch)"
          },
          "finishedAt": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Job completion timestamp (ms since epoch), or null if not finished"
          }
        },
        "required": ["jobId", "state", "createdAt"]
      },
      "PageMetadata": {
        "type": "object",
        "properties": {
          "limit": { "type": "integer" },
          "offset": { "type": "integer" },
          "total": { "type": "integer" },
          "hasNext": { "type": "boolean" },
          "hasPrev": { "type": "boolean" },
          "nextOffset": {
            "type": "integer",
            "nullable": true
          },
          "prevOffset": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": ["limit", "offset", "total", "hasNext", "hasPrev"]
      },
...
      "TechstackAnalyzeRequest": {
        "type": "object",
        "properties": {
          "technologies": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "version": { "type": "string", "nullable": true }
              },
              "required": ["name"],
              "additionalProperties": false
            }
          },
          "waf": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "version": { "type": "string", "nullable": true }
              },
              "required": ["name"],
              "additionalProperties": false
            },
            "default": []
          },
          "secureHeaders": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "header": { "type": "string" },
                "description": { "type": "string", "nullable": true },
                "urls": {
                  "type": "array",
                  "items": { "type": "string" },
                  "default": []
                }
              },
              "required": ["header"],
              "additionalProperties": false
            },
            "default": []
          },
          "cookies": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "domain": { "type": "string", "nullable": true },
                "value": { "type": "string", "nullable": true },
                "secure": { "type": "boolean", "default": false },
                "httpOnly": { "type": "boolean", "default": false },
                "sameSite": { "type": "string", "nullable": true },
                "expirationDate": { "type": "number", "nullable": true }
              },
              "required": ["name"],
              "additionalProperties": false
            },
            "default": []
          },
          "mainDomain": { "type": "string", "nullable": true },
          "raw": {
            "type": "object",
            "description": "Auxiliary raw data blob.",
            "additionalProperties": true
          },
          "storage": {
            "type": "object",
            "description": "Auxiliary storage blob (e.g. localStorage / sessionStorage snapshot).",
            "additionalProperties": true
          }
        },
        "required": ["technologies"],
        "additionalProperties": false
      },
      "TechstackCveSummary": {
        "type": "object",
        "properties": {
          "iri": { "type": "string" },
          "id": { "type": "string" },
          "severity": { "type": "string" },
          "score": { "type": "number" }
        },
        "required": ["iri"]
      },
      "TechstackSoftwareEvidence": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Evidence type: Technology or WAF"
          },
          "name": { "type": "string" },
          "version": { "type": "string" },
          "cpe": {
            "type": "array",
            "items": { "type": "string" }
          },
          "cve": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TechstackCveSummary" }
          }
        }
      },
      "TechstackHeaderEvidence": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "urls": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "TechstackCookieEvidence": {
        "type": "object",
        "properties": {
          "iri": { "type": "string" },
          "name": { "type": "string" },
          "domain": { "type": "string" },
          "path": { "type": "string" },
          "secure": { "type": "boolean" },
          "httpOnly": { "type": "boolean" },
          "sameSite": { "type": "string" },
          "expiration": { "type": "number" }
        },
        "required": ["iri"]
      },
      "TechstackFindingDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "resolver": { "type": "string" },
          "vulnerabilityType": { "type": "string" },
          "severity": { "type": "string" },
          "findingCategory": { "type": "string" },
          "owaspCategory": { "type": "string" },
          "ruleId": { "type": "string" },
          "description": { "type": "string" },
          "remediation": { "type": "string" },
          "evidenceType": { "type": "string" },
          "software": {
            "$ref": "#/components/schemas/TechstackSoftwareEvidence"
          },
          "header": {
            "$ref": "#/components/schemas/TechstackHeaderEvidence"
          },
          "cookies": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TechstackCookieEvidence" }
          }
        },
        "required": ["id"]
      },
      "TechstackFindingListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "type": "string" }
          },
          "page": { "$ref": "#/components/schemas/PageMetadata" }
        },
        "required": ["items", "page"]
      },
...
    },
    "securitySchemes": {}
  }
}